LPS-63538 move
