+++ portal-impl/src/com/liferay/portal/upgrade/Pre7UpgradeProcess.java
@@ -21,10 +21,9 @@ import com.liferay.portal.kernel.bean.PortalBeanLocatorUtil;
 import com.liferay.portal.kernel.model.ResourcePermission;
 import com.liferay.portal.kernel.upgrade.UpgradeException;
 import com.liferay.portal.kernel.upgrade.UpgradeProcess;
-import com.liferay.portal.service.ServiceWrapperProxyUtil;
+import com.liferay.portal.spring.aop.ServiceWrapperProxyUtil;
 
 import java.io.Closeable;
-
 import java.util.List;
 
 /**
